DEF0362546: PRB1631940: MDEV-35217 - Provide information for debugging parallel replication
Implement an option --slave-parallel-print-all-deadlocks to debug
parallel replication conflicts. This option enables additional
information in the server's error log when a parallel replication
conflict or deadlock occurs, similar to --innodb-print-all-deadlocks.
This should be very useful to help investigate the cause of problems
with parallel replication, including helping to find the root cause of
any bugs.

The information printed includes the GTID and thread info of conflicting
threads; the chain of wait_for_prior_commits; and the SHOW ENGINE STATUS of
engines participating in the current transaction.

Setting @@slave_parallel_print_all_deadlocks to 1 will output the
information when an unexpected conflict occurs; this means an event group
that needed to retry more than once, or any lock wait timeout. Setting
@@slave_parallel_print_all_deadlocks to 2 will output the information for
any detected conflict.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>