STRY59256729: DRY mode for CHECK TABLE
Problem
=======

Currently, if CHECK TABLE detects any discrepancies in internal table structures,
it may mark an index or the entire table as corrupted.

As the result, we may not be able to add a new index or even read from the table.

Proposed solution
=================

Add session variable snc_check_table_dry_mode = ON (default) | OFF

When dry mode is ON, CHECK TABLE will conduct the same checks and report findings,
but it will not make any changes to the table.
This feature is implemented for InnoDB storage engine only. Other engines will
execute CHECK TABLE  in a non-safe way, regardless of the variable value.