STRY56031688: Safe replication of ALTER TABLE ddl
Problem
=======
The master executes ALTER TABLE ADD KEY on a temporary table and then drops
it. The ADD KEY operation is scheduled for execution in a separate domain
(domain 7), while the DROP TABLE goes to the default domain (domain 0).

Because domain 7 might be busy (for example, executing another ADD KEY), it
is possible that the DROP TABLE begins before the ADD KEY. As a result, the
slave raises an error - "table does not exist" and stops replication.

Workaround
==========
There is a global variable, slave_ddl_exec_mode, which can be set to STRICT
or IDEMPOTENT (default). In IDEMPOTENT mode, the replica does not stop for
failed DDL operations that would not cause a difference between the primary
and the replica. For example, CREATE TABLE is treated as CREATE OR REPLACE
TABLE, and DROP TABLE is treated as DROP TABLE IF EXISTS.

This change extends the same behavior to ALTER TABLE.
When slave_ddl_exec_mode = IDEMPOTENT, all ALTER TABLE statements on the slave
are treated as ALTER TABLE IF EXISTS.