DEF0596622: PRB1822465: MDEV-36137 - Parallel replication deadlock between DML in one domain and DDL in another
A deadlock could occur with out-of-order parallel replication, when a DDL in
domain B runs in parallel with conflicting DML in domain A. The problem
occurs when one transaction T2 in domain A is blocking the DDL, but is
itself waiting on an earlier transaction T1, and T1 is in turn blocked by
the DDL. This happens because a DDL waiting on an exclusive table lock will
block new transactions from getting even a shared lock.

This patch has a test case that demonstrates this. And a fix for this
specific case, which detects that the DDL is blocked by a later transaction
T2 and deadlock kills T2 (so it will roll back and retry after the DDL
completes).

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>