Refinements to dependent_gtid handling
Replace the mechanism for when eg. DROP TABLE waits for an out-of-order ADD
INDEX running out-of-order in the alternate domain.

No longer uses rpl_global_gtid_waiting.wait_for_pos(); this had the problem
that if the waited-for event group fails, the GTID will not be committed and
the wait will not terminate, so slave stop will hang.

Instead keep a mapping of which dependent_gtid are pending in the alternate
domain, along with their rgi. If the dependent_gtid is not yet committed, do
a wait_for_prior_commit() before running the dependent event group. This
wait mechanims works similar to the normal in-order commit handling, and
will not hang during slave stop, even if the waited-for event group fails.

Also a draft attempt to make DROP TABLE actually insert the dependent_gtid
in the GTID event for a prior ADD INDEX, by opening the table and checking
the TABLE_SHARE::alt_gtid. This makes the future_sync testcase pass, however
other test cases now assert when the table is actually a view (eg. testcase
main.view); more work is needed to correctly handle how the tables (or
views) are opened and closed during DROP, if this method is to be used.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>