Fix parallel replication stalls due to missing wait report
In certain situations, while one transaction was waiting for a row lock,
InnoDB could grant another transaction a lock that conflicted with the waiting
lock, but not with other already granted locks. Then if the page was
reorganised, the granted lock could be moved to the head of the lock queue.
This could end up with a situation where the waiting lock is now waiting for
the granted lock, and this wait was never reported to replication. This could
cause parallel replication to hang for --innodb-lock-wait-timeout if the
waiting transaction is before the granted one.

This patch fixes the issue by adding logic, so that when a lock is granted and
added to a lock queue, a wait is reported against this lock for any already
waiting locks. (This supplements the wait reports made when a waiting lock
is added to the queue).

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>