Experimental: More fine-grained retry in aggressive mode
When a transaction in aggressive parallel replication needs to retry,
identify the conflicting transaction (if any), and wait only for that
one to commit (as opposed to all prior transactions).

This potentially allows more parallelims with larger worker thread
counts and large number of conflicts, at the cost of transactions
possibly requiring multiple rollback/retry.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>