EXPERIMENTAL: --slave-domain-parallel-transactions option
When there is a mix of long- and short-running transactions, optimistic
parallel replication benefits from being able to execute a large amounts of
short transactions that follow a long one, to keep CPU cores busy. This
requires a large value of --slave-parallel-threads, since each transaction
waiting for prior commit consumes one worker thread.

But starting a large number of transactions in parallel can be sub-optimal.
The OS may schedule the threads in a bad order (running later transactions
before earlier ones); and too many running threads can lead to excessive
conflicts and synchronisation overhead.

This new --slave-domain-parallel-transactions option allows to limit the
number of transactions that are started in parallel. This way, we can
configure optimistic parallel replication to start say 16 transactions in
parallel, but allow up to say 256 transactions to be pending at the same
time. Each time a transaction gets ready to commit and is about to
wait_for_prior_commit(), it will signal another transaction to be allowed to
start. This (soft) limits the amount of active threads to 16, while still
allowing up to 255 small transactions to replicate in parallel with a
long-running one.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>