Option to avoid binlogging queries that do not modify any rows
In mixed binlogging format, when a query is binlogged as a statement, that
statement gets binlogged even if the query does not update any rows. This is
unlike when a query gets binlogged in row mode, in which case nothing is
binlogged when no rows are modified.

Add an option to omit the binlogging of such queries also when
statement-based binlogging would otherwise be used.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>